Combine ParsedRuntimeEnv and RuntimeEnv into ray.runtime.RuntimeEnv
