+++ python/ray/_private/runtime_env/validation.py
@@ -1,22 +1,12 @@
 import logging
-import os
 from pathlib import Path
 import sys
-from typing import Any, Dict, List, Optional, Set, Union
+from typing import Dict, List, Optional, Union
 from pkg_resources import Requirement
 from collections import OrderedDict
 import yaml
 
-import ray
-from ray._private.runtime_env.plugin import RuntimeEnvPlugin, encode_plugin_uri
-from ray._private.runtime_env.utils import RuntimeEnv
-from ray._private.utils import import_attr
-from ray._private.runtime_env.conda import (
-    _resolve_install_from_source_ray_extras,
-    get_uri as get_conda_uri,
-)
-
-from ray._private.runtime_env.pip import get_uri as get_pip_uri
+from ray._private.runtime_env.conda import _resolve_install_from_source_ray_extras
 
 logger = logging.getLogger(__name__)
 
@@ -272,177 +262,3 @@ OPTION_TO_VALIDATION_FN = {
     "env_vars": parse_and_validate_env_vars,
     "container": parse_and_validate_container,
 }
-
-
-class ParsedRuntimeEnv(dict):
-    """An internal wrapper for runtime_env that is parsed and validated.
-
-    This should be constructed from user-provided input (the API runtime_env)
-    and used everywhere that the runtime_env is passed around internally.
-
-    All options in the resulting dictionary will have non-None values.
-
-    Currently supported options:
-        py_modules (List[URI]): List of URIs (either in the GCS or external
-            storage), each of which is a zip file that will be unpacked and
-            inserted into the PYTHONPATH of the workers.
-        working_dir (URI): URI (either in the GCS or external storage) of a zip
-            file that will be unpacked in the directory of each task/actor.
-        pip (List[str] | str): Either a list of pip packages, or a string
-            containing the path to a pip requirements.txt file.
-        conda (dict | str): Either the conda YAML config, the name of a
-            local conda env (e.g., "pytorch_p36"), or the path to a conda
-            environment.yaml file.
-            The Ray dependency will be automatically injected into the conda
-            env to ensure compatibility with the cluster Ray. The conda name
-            may be mangled automatically to avoid conflicts between runtime
-            envs.
-            This field cannot be specified at the same time as the 'pip' field.
-            To use pip with conda, please specify your pip dependencies within
-            the conda YAML config:
-            https://conda.io/projects/conda/en/latest/user-guide/tasks/manage-e
-            nvironments.html#create-env-file-manually
-            Examples:
-                {"channels": ["defaults"], "dependencies": ["codecov"]}
-                "pytorch_p36"   # Found on DLAMIs
-        container (dict): Require a given (Docker) container image,
-            The Ray worker process will run in a container with this image.
-            The `worker_path` is the default_worker.py path.
-            The `run_options` list spec is here:
-            https://docs.docker.com/engine/reference/run/
-            Examples:
-                {"image": "anyscale/ray-ml:nightly-py38-cpu",
-                 "worker_path": "/root/python/ray/workers/default_worker.py",
-                 "run_options": ["--cap-drop SYS_ADMIN","--log-level=debug"]}
-        env_vars (dict): Environment variables to set.
-            Examples:
-                {"OMP_NUM_THREADS": "32", "TF_WARNINGS": "none"}
-    """
-
-    known_fields: Set[str] = {
-        "py_modules",
-        "working_dir",
-        "conda",
-        "pip",
-        "container",
-        "excludes",
-        "env_vars",
-        "_ray_release",
-        "_ray_commit",
-        "_inject_current_ray",
-        "plugins",
-        "eager_install",
-    }
-
-    def __init__(self, runtime_env: Dict[str, Any], _validate: bool = True):
-        super().__init__()
-        self._cached_pb = None
-
-        # Blindly trust that the runtime_env has already been validated.
-        # This is dangerous and should only be used internally (e.g., on the
-        # deserialization codepath.
-        if not _validate:
-            self.update(runtime_env)
-            return
-
-        if runtime_env.get("conda") and runtime_env.get("pip"):
-            raise ValueError(
-                "The 'pip' field and 'conda' field of "
-                "runtime_env cannot both be specified.\n"
-                f"specified pip field: {runtime_env['pip']}\n"
-                f"specified conda field: {runtime_env['conda']}\n"
-                "To use pip with conda, please only set the 'conda' "
-                "field, and specify your pip dependencies "
-                "within the conda YAML config dict: see "
-                "https://conda.io/projects/conda/en/latest/"
-                "user-guide/tasks/manage-environments.html"
-                "#create-env-file-manually"
-            )
-
-        for option, validate_fn in OPTION_TO_VALIDATION_FN.items():
-            option_val = runtime_env.get(option)
-            if option_val is not None:
-                validated_option_val = validate_fn(option_val)
-                if validated_option_val is not None:
-                    self[option] = validated_option_val
-
-        if "_ray_release" in runtime_env:
-            self["_ray_release"] = runtime_env["_ray_release"]
-
-        if "_ray_commit" in runtime_env:
-            self["_ray_commit"] = runtime_env["_ray_commit"]
-        else:
-            if self.get("pip") or self.get("conda"):
-                self["_ray_commit"] = ray.__commit__
-
-        # Used for testing wheels that have not yet been merged into master.
-        # If this is set to True, then we do not inject Ray into the conda
-        # or pip dependencies.
-        if "_inject_current_ray" in runtime_env:
-            self["_inject_current_ray"] = runtime_env["_inject_current_ray"]
-        elif "RAY_RUNTIME_ENV_LOCAL_DEV_MODE" in os.environ:
-            self["_inject_current_ray"] = True
-
-        if "plugins" in runtime_env:
-            self["plugins"] = dict()
-            for class_path, plugin_field in runtime_env["plugins"].items():
-                plugin_class: RuntimeEnvPlugin = import_attr(class_path)
-                if not issubclass(plugin_class, RuntimeEnvPlugin):
-                    # TODO(simon): move the inferface to public once ready.
-                    raise TypeError(
-                        f"{class_path} must be inherit from "
-                        "ray._private.runtime_env.plugin.RuntimeEnvPlugin."
-                    )
-                # TODO(simon): implement uri support.
-                _ = plugin_class.validate(runtime_env)
-                # Validation passed, add the entry to parsed runtime env.
-                self["plugins"][class_path] = plugin_field
-
-        unknown_fields = set(runtime_env.keys()) - ParsedRuntimeEnv.known_fields
-        if len(unknown_fields):
-            logger.warning(
-                "The following unknown entries in the runtime_env dictionary "
-                f"will be ignored: {unknown_fields}. If you intended to use "
-                "them as plugins, they must be nested in the `plugins` field."
-            )
-
-        # NOTE(architkulkarni): This allows worker caching code in C++ to check
-        # if a runtime env is empty without deserializing it.  This is a catch-
-        # all; for validated inputs we won't set the key if the value is None.
-        if all(val is None for val in self.values()):
-            self.clear()
-
-    def get_uris(self) -> List[str]:
-        # TODO(architkulkarni): this should programmatically be extended with
-        # URIs from all plugins.
-        plugin_uris = []
-        if "working_dir" in self:
-            plugin_uris.append(encode_plugin_uri("working_dir", self["working_dir"]))
-        if "py_modules" in self:
-            for uri in self["py_modules"]:
-                plugin_uris.append(encode_plugin_uri("py_modules", uri))
-        if "conda" in self:
-            uri = get_conda_uri(self)
-            if uri is not None:
-                plugin_uris.append(encode_plugin_uri("conda", uri))
-        if "pip" in self:
-            uri = get_pip_uri(self)
-            if uri is not None:
-                plugin_uris.append(encode_plugin_uri("pip", uri))
-
-        return plugin_uris
-
-    def get_proto_runtime_env(self):
-        """Return the protobuf structure of runtime env."""
-        if self._cached_pb is None:
-            self._cached_pb = RuntimeEnv.from_dict(self, get_conda_uri, get_pip_uri)
-
-        return self._cached_pb
-
-    @classmethod
-    def deserialize(cls, serialized: str) -> "ParsedRuntimeEnv":
-        runtime_env = RuntimeEnv(serialized_runtime_env=serialized)
-        return cls(runtime_env.to_dict(), _validate=False)
-
-    def serialize(self) -> str:
-        return self.get_proto_runtime_env().serialize()
