+++ python/ray/remote_function.py
@@ -14,7 +14,7 @@ from ray._private.client_mode_hook import client_mode_convert_function
 from ray._private.client_mode_hook import client_mode_should_convert
 from ray.util.placement_group import configure_placement_group_based_on_context
 import ray._private.signature
-from ray._private.runtime_env.validation import ParsedRuntimeEnv
+from ray.runtime_env import RuntimeEnv
 from ray.util.tracing.tracing_helper import (
     _tracing_task_invocation,
     _inject_tracing_into_function,
@@ -146,7 +146,7 @@ class RemoteFunction:
             if isinstance(runtime_env, str):
                 self._runtime_env = runtime_env
             else:
-                self._runtime_env = ParsedRuntimeEnv(runtime_env or {}).serialize()
+                self._runtime_env = RuntimeEnv(**(runtime_env or {})).serialize()
         else:
             self._runtime_env = None
         self._placement_group = placement_group
@@ -219,7 +219,7 @@ class RemoteFunction:
                 # Serialzed protobuf runtime env from Ray client.
                 new_runtime_env = runtime_env
             else:
-                new_runtime_env = ParsedRuntimeEnv(runtime_env).serialize()
+                new_runtime_env = RuntimeEnv(**runtime_env).serialize()
         else:
             # Keep the runtime_env as None.  In .remote(), we need to know if
             # runtime_env is None to know whether or not to fall back to the
