+++ python/ray/tests/test_advanced_3.py
@@ -27,7 +27,7 @@ from ray._private.test_utils import (
     wait_for_condition,
     wait_for_num_actors,
 )
-from ray._private.runtime_env.utils import RuntimeEnv
+from ray.runtime_env import RuntimeEnv
 
 logger = logging.getLogger(__name__)
 
@@ -769,8 +769,8 @@ def test_sync_job_config(shutdown_only):
     # Check that the job config is synchronized at the driver side.
     job_config = ray.worker.global_worker.core_worker.get_job_config()
     assert job_config.num_java_workers_per_process == num_java_workers_per_process
-    job_runtime_env = RuntimeEnv(
-        serialized_runtime_env=job_config.runtime_env_info.serialized_runtime_env
+    job_runtime_env = RuntimeEnv.deserialize(
+        job_config.runtime_env_info.serialized_runtime_env
     )
     assert job_runtime_env.env_vars() == runtime_env["env_vars"]
 
@@ -783,8 +783,8 @@ def test_sync_job_config(shutdown_only):
     job_config = gcs_utils.JobConfig()
     job_config.ParseFromString(ray.get(get_job_config.remote()))
     assert job_config.num_java_workers_per_process == num_java_workers_per_process
-    job_runtime_env = RuntimeEnv(
-        serialized_runtime_env=job_config.runtime_env_info.serialized_runtime_env
+    job_runtime_env = RuntimeEnv.deserialize(
+        job_config.runtime_env_info.serialized_runtime_env
     )
     assert job_runtime_env.env_vars() == runtime_env["env_vars"]
