+++ python/ray/tests/test_runtime_env_conda_and_pip.py
@@ -3,7 +3,6 @@ import pytest
 import sys
 import platform
 import time
-from ray._private.runtime_env.utils import RuntimeEnv
 from ray._private.test_utils import (
     wait_for_condition,
     chdir,
@@ -11,10 +10,8 @@ from ray._private.test_utils import (
     generate_runtime_env_dict,
 )
 from ray._private.runtime_env.conda import _get_conda_dict_with_ray_inserted
-from ray._private.runtime_env.validation import (
-    ParsedRuntimeEnv,
-    _rewrite_pip_list_ray_libraries,
-)
+from ray._private.runtime_env.validation import _rewrite_pip_list_ray_libraries
+from ray.runtime_env import RuntimeEnv
 
 import yaml
 import tempfile
@@ -51,7 +48,7 @@ def test_get_conda_dict_with_ray_inserted_m1_wheel(monkeypatch):
     monkeypatch.setattr(platform, "machine", lambda: "arm64")
 
     input_conda = {"dependencies": ["blah", "pip", {"pip": ["pip_pkg"]}]}
-    runtime_env = RuntimeEnv(ParsedRuntimeEnv({"conda": input_conda}).serialize())
+    runtime_env = RuntimeEnv(conda=input_conda).serialize()
     output_conda = _get_conda_dict_with_ray_inserted(runtime_env)
     # M1 wheels are not uploaded to AWS S3.  So rather than have an S3 URL
     # inserted as a dependency, we should just have the string "ray==1.9.0".
